scheduler/requirements.txt
pandas==1.3.3
numpy==1.21.2
python-dotenv==1.0.1
requests==2.32.3